simulation_data/mqtt/requirements.txt
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
PyYAML>=6.0.1
